backend/requirements.txt
# FastAPI & Web
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pgvector==0.2.4

# AI & ML
openai>=1.12.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
chromadb>=0.4.22
tiktoken>=0.5.2

# Document Processing
PyMuPDF==1.23.22
python-docx==1.1.0
Pillow==10.2.0

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
aiofiles==23.2.1
tenacity==8.2.3

# Development
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
black==24.1.1
ruff==0.2.0
